kafka/consumer/requirements.txt
# kafka/consumer/requirements.txt
kafka-python
psycopg2-binary